github/pull/extracts/EagleAI-Research-Trades-ccbde6bf15deb241d86c643bbaa7bc31db9f8672/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
apexomni==3.0.8
appnope==0.1.4
APScheduler==3.11.0
asttokens==3.0.1
attrs==25.3.0
bidict==0.23.1
binance-common==3.5.0
binance-sdk-derivatives-trading-usds-futures==7.0.0
bitarray==3.7.1
black==25.12.0
blinker==1.9.0
cachetools==6.2.2
certifi==2025.8.3
charset-normalizer==3.4.3
ckzg==2.1.4
click==8.3.0
comm==0.2.3
cytoolz==1.0.1
debugpy==1.8.19
decorator==5.2.1
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
executing==2.2.1
flasgger==0.9.7.1
Flask==3.1.2
flask-cors==6.0.1
frozenlist==1.7.0
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-pubsub==2.33.0
googleapis-common-protos==1.72.0
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
hexbytes==1.3.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.7.0
jupyter_core==5.9.1
MarkupSafe==3.0.2
matplotlib-inline==0.2.1
mistune==3.1.4
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
nest-asyncio==1.6.0
netifaces==0.10.6
numpy==2.3.3
openai==2.2.0
opentelemetry-api==1.39.0
opentelemetry-sdk==1.39.0
opentelemetry-semantic-conventions==0.60b0
packaging==25.0
parsimonious==0.10.0
parso==0.8.5
pathspec==1.0.4
pexpect==4.9.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.33.2
psutil==7.1.3
psycopg2==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycryptodome==3.23.0
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.15.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.13.0
python-socketio==5.16.0
pytokens==0.4.1
pyunormalize==16.0.0
PyYAML==6.0.3
pyzmq==27.1.0
redis==7.1.0
referencing==0.37.0
regex==2025.9.18
requests==2.32.5
rlp==4.1.0
rpds-py==0.28.0
rsa==4.9.1
ruff==0.12.12
setuptools==3.3
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
socketio==0.2.1
stack-data==0.6.3
sympy==1.10.1
toolz==1.0.0
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
types-requests==2.32.4.20250913
typing-inspection==0.4.1
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.5.0
wcwidth==0.2.14
web3==7.13.0
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.3.2
yarl==1.20.1
zipp==3.23.0
